mv profcs specific struct to linux.rs, make undefined ports sortable, re/move comments
